registry: move module logic into its own files
