+++ internal/registry/module_mock_responses_test.go
