+++ internal/registry/module_test.go
